Uploaded the first stable version.
This version has a number of items that need to be addressed, such as ensuring that a proper date range is entered, as well as plenty of data formatting issues.  It was just supposed to be a quick project to show off many of the features I learned in the course.  In later versions, I'd like to incorporate Unit Testing, Visualization using Bokah, and other User Interface options, such as Search.